requirements/base.txt
numpy
pandas
pillow
scikit-learn
matplotlib
tqdm
PyYAML
opencv-python
kagglehub
yt_dlp
